helper functions (plots, postproc, timer)

+++ tools/postprocessing.py
@@ -21,4 +21,4 @@ def classify(mask: np.ndarray) -> Tuple[np.ndarray, bool]:
 
 @tf.function
 def tf_classify(output: tf.Tensor) -> tf.Tensor:
-    return tf.cast(tf.greater(output, tf.constant(0.5)), tf.uint8)
+    return tf.cast(tf.greater(output, tf.constant(0.5)), tf.int64)
